total cost, checkout button configure

+++ NumNum/Food/Scenes/Basket/View/BasketController.swift
@@ -13,37 +13,63 @@ class BasketController: UIViewController {
     @IBOutlet  weak var collection: UICollectionView!
     @IBOutlet weak var cartTotalLabel: UILabel!
     
+    @IBOutlet weak var payView: UIView!
+    @IBOutlet weak var checkoutButton: UIButton!
+    
+    
+    var isSelected: Bool = true
+    
     var basketViewModel = BasketViewModel()
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
-        collectionSetup()
-        basketViewModel.getDataFromFirebase()
-        isSucceed()
+        viewDidLoad()
+        
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        collectionSetup()
+        basketViewModel.getDataFromFirebase()
+        isSucceed()
+        configureUI()
         
     }
     
+    //MARK: - ConfigureUI Buttons
+
+    func configureUI() {
+        payView.layer.cornerRadius = 25
+        checkoutButton.layer.cornerRadius = 10
+    }
+    
+    //MARK: - Button actions
+
     @IBAction func checkoutBtnTapped(_ sender: UIButton) {
-        
+        basketViewModel.updateFirebaseData()
+        basketViewModel.getDataFromFirebase()
+        isSucceed()
+        if basketViewModel.dishes.count > 0 {
+            DuplicateFuncs.alertMessage(title: "Your order has been taken!", message: "", vc: self)
+        } else {
+            DuplicateFuncs.alertMessage(title: "Your cart is empty.", message: "Please add items to your cart.", vc: self)
+        }
     }
     
     //MARK: - Register UINib's
-
+    
     private func collectionSetup() {
         collection.register(UINib(nibName: "\(BasketViewCell.self)", bundle: nil), forCellWithReuseIdentifier: "\(BasketViewCell.self)")
     }
     
     //MARK: - ViewModelConfiguration
-
+    
     private func isSucceed() {
         self.basketViewModel.successCallback = { [ weak self ] in
             self?.collection.reloadData()
+            self?.cartTotalLabel.text = "\(self?.basketViewModel.totalCost ?? 0) Items"
         }
-       
+        
     }
     
 }
@@ -51,7 +77,7 @@ class BasketController: UIViewController {
 //MARK: - CollectionViewMethods
 
 extension BasketController: UICollectionViewDataSource, UICollectionViewDelegate {
-   
+    
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         return basketViewModel.dishes.count
     }
@@ -60,23 +86,36 @@ extension BasketController: UICollectionViewDataSource, UICollectionViewDelegate
         let cell = collection.dequeueReusableCell(withReuseIdentifier: "\(BasketViewCell.self)", for: indexPath) as! BasketViewCell
         cell.getDataFromFirebase(data: basketViewModel.dishes[indexPath.row])
         cell.configure(data: basketViewModel.dishes[indexPath.row])
+        cell.layer.cornerRadius = 10
         cell.delegate = self
         return cell
         
+        
     }
-
-
+    
+    
 }
 
 //MARK: - BasketViewCellSubclassDelegate
 
 extension BasketController: BasketViewCellSubclassDelegate {
-    func stepperValuDidChange(cell: BasketViewCell) {
+    func stepperValueDidChange(cell: BasketViewCell, stepper: UIStepper) {
         guard let indexPath = self.collection.indexPath(for: cell) else {
             return
         }
-        basketViewModel.dishes.remove(at: indexPath.row)
-        collection.deleteItems(at: [indexPath])
+        
+        if stepper.value == 0 {
+            basketViewModel.getDataFromFirebase()
+            isSucceed()
+            basketViewModel.dishes.remove(at: indexPath.row)
+            collection.deleteItems(at: [indexPath])
+            
+        } else {
+            basketViewModel.getDataFromFirebase()
+            isSucceed()
+            
+        }
+        
     }
 }
